+++ CHANGELOG.md
@@ -3,12 +3,16 @@
 ## Added
 
 * Adding `details` from `ClosePollOp` to `creator_details` of polls after its closing
+* filter `asset_owner` for balances list
 * Endpoint `/v3/votes/{voter}` which returns all the _votes_ created by specific _voter_ with relationships `Poll` (_poll_ where the _vote_ was created) and `Account` (_voter_ account)
 
 ### Fixed
 
 * Sales not allowed to participate in appearing on`/v3/account/{id}/sales`
 * Transaction failure response for `/v3/transactions`
+* Waiting for transaction ingestion for `/v3/transactions`
+* request details for create atomic swap ask and bid requests
+* response on `v3/create_atomic_swap_aks_requests` (quote assets include)
 
 ## 3.4.0-x.0